revert migration of enums from SurfaceNamer to Enums
