+++ src/main/java/com/google/api/codegen/discogapic/transformer/DiscoGapicNamer.java
@@ -16,7 +16,6 @@ package com.google.api.codegen.discogapic.transformer;
 
 import com.google.api.codegen.discovery.Method;
 import com.google.api.codegen.transformer.SurfaceNamer;
-import com.google.api.codegen.util.Enums;
 import com.google.api.codegen.util.Name;
 import com.google.api.codegen.util.TypeName;
 import com.google.api.codegen.util.TypeNameConverter;
@@ -35,7 +34,7 @@ public class DiscoGapicNamer {
 
   /** Returns the resource setter method name for a resource field. */
   public String getResourceSetterName(
-      String fieldName, Enums.Cardinality isRepeated, SurfaceNamer languageNamer) {
+      String fieldName, SurfaceNamer.Cardinality isRepeated, SurfaceNamer languageNamer) {
     switch (isRepeated) {
       case IS_REPEATED:
         return languageNamer.publicMethodName(
